Use puppetlabs/xinetd for cups-lpd
Hands off responsibility for defining and managing the xinetd service
to puppetlabs/xinetd module.  Avoids service resource duplicate errors
for the xinetd service.

Use of xinetd::service in config causes Service[‘xinetd’] to be defined
and receive notifications of changes in the service file, so it should
not need to be included in cups::service.